refactor: Connection and TransformationRule as Model instead of Message
Even if connections and transformation rules are sent from go to python,
they are also models backed by a tool table.
