+++ backend/python/pydevlake/pydevlake/__init__.py
@@ -14,9 +14,9 @@
 # limitations under the License.
 
 
-from .model import ToolModel, ToolScope, DomainScope
+from .model import ToolModel, ToolScope, DomainScope, Connection, TransformationRule
 from .logger import logger
-from .message import Connection, TransformationRule, RemoteScope
+from .message import RemoteScope
 from .plugin import Plugin
 from .stream import DomainType, Stream, Substream
 from .context import Context
