+++ backend/python/pydevlake/pydevlake/model.py
@@ -28,17 +28,29 @@ import inflect
 inflect_engine = inflect.engine()
 
 
-def get_plugin_name(cls):
-    """
-    Get the plugin name from a class by looking into
-    the file path of its module.
-    """
-    module = getmodule(cls)
-    path_segments = module.__file__.split(os.sep)
-    # Finds the name of the first enclosing folder
-    # that is not a python module 
-    depth = len(module.__name__.split('.')) + 1
-    return path_segments[-depth]
+class Model(SQLModel):
+    id: int = Field(primary_key=True)
+    created_at: Optional[datetime] = Field(
+        sa_column=Column(DateTime(), default=func.now())
+    )
+    updated_at: Optional[datetime] = Field(
+        sa_column=Column(DateTime(), default=func.now(), onupdate=func.now())
+    )
+
+class ToolTable(Model):
+    @declared_attr
+    def __tablename__(cls) -> str:
+        plugin_name = _get_plugin_name(cls)
+        plural_entity = inflect_engine.plural_noun(cls.__name__.lower())
+        return f'_tool_{plugin_name}_{plural_entity}'
+
+
+class Connection(ToolTable):
+    name: str
+
+
+class TransformationRule(ToolTable):
+    name: str
 
 
 class RawModel(SQLModel):
@@ -64,20 +76,7 @@ class RawDataOrigin(SQLModel):
         self.raw_data_table = raw.__tablename__
 
 
-class ToolModel(RawDataOrigin):
-    @declared_attr
-    def __tablename__(cls) -> str:
-        plugin_name = get_plugin_name(cls)
-        plural_entity = inflect_engine.plural_noun(cls.__name__.lower())
-        return f'_tool_{plugin_name}_{plural_entity}'
-
-
-class ToolScope(ToolModel):
-    id: str = Field(primary_key=True)
-    name: str
-
-
-class NoPKModel(SQLModel):
+class NoPKModel(RawDataOrigin):
     created_at: datetime = Field(
         sa_column=Column(DateTime(), default=func.now())
     )
@@ -86,8 +85,21 @@ class NoPKModel(SQLModel):
     )
 
 
+class ToolModel(ToolTable, NoPKModel):
+    @declared_attr
+    def __tablename__(cls) -> str:
+        plugin_name = _get_plugin_name(cls)
+        plural_entity = inflect_engine.plural_noun(cls.__name__.lower())
+        return f'_tool_{plugin_name}_{plural_entity}'
+
+
 class DomainModel(NoPKModel):
     id: str = Field(primary_key=True)
+
+
+class ToolScope(ToolModel):
+    id: str = Field(primary_key=True)
+    name: str
 
 
 class DomainScope(DomainModel):
@@ -100,10 +112,23 @@ def generate_domain_id(tool_model: ToolModel, connection_id: str):
     from the tool entity it originates from.
     """
     model_type = type(tool_model)
-    segments = [get_plugin_name(model_type), model_type.__name__, str(connection_id)]
+    segments = [_get_plugin_name(model_type), model_type.__name__, str(connection_id)]
     mapper = inspect(model_type)
     for primary_key_column in mapper.primary_key:
         prop = mapper.get_property_by_column(primary_key_column)
         attr_val = getattr(tool_model, prop.key)
         segments.append(str(attr_val))
     return ':'.join(segments)
+
+
+def _get_plugin_name(cls):
+    """
+    Get the plugin name from a class by looking into
+    the file path of its module.
+    """
+    module = getmodule(cls)
+    path_segments = module.__file__.split(os.sep)
+    # Finds the name of the first enclosing folder
+    # that is not a python module
+    depth = len(module.__name__.split('.')) + 1
+    return path_segments[-depth]
